requirements.txt
fastapi>=0.110
pydantic>=2.6
pyyaml>=6.0
uvicorn>=0.29
